extracted public method to show warning dialog and disable plugin
