+++ platform/platform-impl/src/com/intellij/diagnostic/IdeErrorsDialog.java
@@ -28,7 +28,6 @@ import com.intellij.ide.plugins.PluginManagerMain;
 import com.intellij.ide.plugins.cl.PluginClassLoader;
 import com.intellij.ide.util.PropertiesComponent;
 import com.intellij.openapi.actionSystem.*;
-import com.intellij.openapi.application.Application;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.application.ApplicationNamesInfo;
 import com.intellij.openapi.application.ex.ApplicationEx;
@@ -47,7 +46,6 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.DialogWrapper;
 import com.intellij.openapi.util.Comparing;
 import com.intellij.openapi.util.Condition;
-import com.intellij.openapi.util.Ref;
 import com.intellij.openapi.util.SystemInfo;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.openapi.wm.IdeFocusManager;
@@ -56,7 +54,6 @@ import com.intellij.ui.HeaderlessTabbedPane;
 import com.intellij.ui.HyperlinkLabel;
 import com.intellij.ui.IdeBorderFactory;
 import com.intellij.util.Consumer;
-import com.intellij.util.Function;
 import com.intellij.util.containers.ContainerUtil;
 import com.intellij.util.text.DateFormatUtil;
 import com.intellij.util.ui.UIUtil;
@@ -401,43 +398,8 @@ public class IdeErrorsDialog extends DialogWrapper implements MessagePoolListene
 
   private void disablePlugin() {
     final PluginId pluginId = findPluginId(getSelectedMessage().getThrowable());
-    if (pluginId == null) {
-      return;
-    }
-
-    IdeaPluginDescriptor plugin = PluginManager.getPlugin(pluginId);
-    final Ref<Boolean> hasDependants = new Ref<Boolean>(false);
-    PluginManager.checkDependants(plugin, new Function<PluginId, IdeaPluginDescriptor>() {
-                                    @Override
-                                    public IdeaPluginDescriptor fun(PluginId pluginId) {
-                                      return PluginManager.getPlugin(pluginId);
-                                    }
-                                  }, new Condition<PluginId>() {
-      @Override
-      public boolean value(PluginId pluginId) {
-        if (PluginManagerCore.CORE_PLUGIN_ID.equals(pluginId.getIdString())) {
-          return true;
-        }
-        hasDependants.set(true);
-        return false;
-      }
-    }
-    );
-
-    Application app = ApplicationManager.getApplication();
-    DisablePluginWarningDialog d =
-      new DisablePluginWarningDialog(getRootPane(), plugin.getName(), hasDependants.get(), app.isRestartCapable());
-    d.show();
-    switch (d.getExitCode()) {
-      case CANCEL_EXIT_CODE:
-        return;
-      case DisablePluginWarningDialog.DISABLE_EXIT_CODE:
-        PluginManager.disablePlugin(pluginId.getIdString());
-        break;
-      case DisablePluginWarningDialog.DISABLE_AND_RESTART_EXIT_CODE:
-        PluginManager.disablePlugin(pluginId.getIdString());
-        app.restart();
-        break;
+    if (pluginId != null) {
+      DisablePluginWarningDialog.disablePlugin(pluginId, getRootPane());
     }
   }
